enable hmr for welcome page
