--- superset/assets/src/welcome/index.jsx
@@ -1,22 +1,0 @@
-/* eslint no-unused-vars: 0 */
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Panel, Row, Col, FormControl } from 'react-bootstrap';
-
-import { appSetup } from '../common';
-import App from './App';
-
-appSetup();
-
-const container = document.getElementById('app');
-const bootstrap = JSON.parse(container.getAttribute('data-bootstrap'));
-const user = {
-  ...bootstrap.user,
-};
-
-ReactDOM.render(
-  <App
-    user={user}
-  />,
-  container,
-);
